+++ src/main/java/wooteco/subway/line/controller/dto/SectionRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.line.dto;
+package wooteco.subway.line.controller.dto;
 
 public class SectionRequest {
     private Long upStationId;
