+++ src/main/java/wooteco/subway/path/controller/PathController.java
@@ -1,4 +1,4 @@
-package wooteco.subway.path.ui;
+package wooteco.subway.path.controller;
 
 public class PathController {
     // TODO: 경로조회 기능 구현하기
